[Job]Do not poll Job status with RequeueAfter
Polling with RequeueAfter is not necessary as operators should be
watchign the Jobs they create and therefore automatically get reconciled
when the Job status changes.

This change removes all RequeueAfter in the Job handling code path. We
still have to provide information to the caller about the status of
the status of the Job so the caller then decide when can move forward
with other tasks depending on the success of the Job. So DoJob is
changed to return an enum instead of a ctrl.Result to better
differentiate between Running/Failed/Succeeded Job statuses.

Note that we had and still have an implicit contract between the caller
via the Job hash. If the current Job's hash is not different from the
hash the caller has persisted then:
- no new Job is needed
- if there is no Job exists it means that the Job succeeded in the past
  and got cleaned up.